Attempt to fix cirrus fail
