+++ app/src/androidTest/java/com/github/h3lp3rs/h3lp/GuideTest.kt
@@ -23,41 +23,5 @@ class GuideTest {
         MainPageActivity::class.java
 
     )
-    private val targetContext: Context = ApplicationProvider.getApplicationContext()
-
-
-    private fun clearPreferences() {
-        val sharedPrefs = PreferenceManager.getDefaultSharedPreferences(targetContext)
-        val editor = sharedPrefs.edit()
-        editor.clear()
-        editor.apply()
-    }
-
-    @Test
-    fun checkThatGuideIsInitiallyNotLaunched() {
-        clearPreferences()
-        val prefManager = PreferenceManager.getDefaultSharedPreferences(targetContext)
-        assertFalse(prefManager.getBoolean("didShowGuide", false))
-    }
-
-    @Test
-    fun checkThatGuideIsLaunched() {
-        val prefManager = PreferenceManager.getDefaultSharedPreferences(targetContext)
-        assertTrue(prefManager.getBoolean("didShowGuide", false))
-    }
-
-    @Test
-    fun finishingAppDemoDisplaysMessage() {
-        clearPreferences()
-        var i = 0
-        // +1 for the search bar
-        val nbButtons = mainPageButtons.size + scrollViewButtons.size + 1
-        while (i < nbButtons) {
-            onView(withId(R.id.HelloText)).perform(click())
-            i++
-        }
-        onView(ViewMatchers.withText(R.string.AppGuideFinished))
-            .check(ViewAssertions.matches(ViewMatchers.withEffectiveVisibility(ViewMatchers.Visibility.VISIBLE)))
-    }
 
 }
